+++ magik-language-server/src/test/java/nl/ramsolutions/sw/magik/languageserver/references/ReferencesProviderTest.java
@@ -66,7 +66,7 @@ class ReferencesProviderTest {
   }
 
   @Test
-  void testProvideMethodReferenceFromMethodDefintion() {
+  void testProvideMethodReferenceFromMethodDefinition() {
     final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
     definitionKeeper.add(
         new MethodDefinition(